% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to the SQLite database}
\usage{
connect_db(path = ".")
}
\arguments{
\item{path}{the path to the database}
}
\description{
The database is an SQLite database named \code{soundcluster.sqlite}. This file will be created if it doesn't exist at the \code{path}.
}
