% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_boundary.R
\name{extract_boundary_pulse}
\alias{extract_boundary_pulse}
\title{Extract boundary points for each pulse}
\usage{
extract_boundary_pulse(spectrogram, threshold_amplitude = 10,
  min_peak_amplitude = 30, amplitude_step = 5)
}
\arguments{
\item{spectrogram}{a \code{soundSpectrogram} object}

\item{threshold_amplitude}{relevant regions have an amplitude above the \code{threshold_amplitude}. Defaults to 10 dB.}

\item{min_peak_amplitude}{the maximum amplitude in a relevant region must be above \code{min_peak_amplitude}. Defaults to 30 dB.}

\item{amplitude_step}{the step size in which the threshold amplitude is increased. Defaults to 5 dB.}
}
\description{
Extract boundary points for each pulse
}
