% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_full_pulse.R
\name{extract_full_pulse}
\alias{extract_full_pulse}
\title{Extract full pulses of a spectrogram}
\usage{
extract_full_pulse(spectrogram, threshold_amplitude = 10,
  min_peak_amplitude = 30, dimensions = 32)
}
\arguments{
\item{spectrogram}{a \code{soundSpectrogram} object}

\item{threshold_amplitude}{relevant regions have an amplitude above the \code{threshold_amplitude}. Defaults to 10 dB.}

\item{min_peak_amplitude}{the maximum amplitude in a relevant region must be above \code{min_peak_amplitude}. Defaults to 30 dB.}

\item{dimensions}{the number of rows and columns used to resample the shape. Must be a single number and a power of 2. Will be altered to the next power of 2.}
}
\description{
Extract full pulses of a spectrogram
}
\examples{
wav <- sound_wav(
  system.file("demo.wav", package = "soundcluster"),
  te_factor = 10,
  max_length = 0.1
)
spectrogram <- wav2spectrogram(wav)
extract_full_pulse(spectrogram, min_peak_amplitude = 20, dimensions = 16)
}
