% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_node_classification.R
\name{get_node_classification}
\alias{get_node_classification}
\title{get the classification for each node in a model}
\usage{
get_node_classification(x, model)
}
\arguments{
\item{x}{a soundDatabase}

\item{model}{an optional model id}
}
\value{
a dataframe with the classification
}
\description{
get the classification for each node in a model
}
