% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_soundCluster.R
\name{plot.soundCluster}
\alias{plot.soundCluster}
\title{plot a soundCluster object}
\usage{
\method{plot}{soundCluster}(x, y, ...)
}
\arguments{
\item{x}{a soundCluster object}

\item{y}{currently ignored}

\item{...}{currently ignored}
}
\description{
Displays the reconstructed average sound for each node
}
