% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_pyramide.R
\name{raw_pyramide}
\alias{raw_pyramide}
\title{Prepare a matrix of unscaled pulse parameters based on generic pulse properties and the pyramides of the shape}
\usage{
raw_pyramide(sound_pulse)
}
\arguments{
\item{sound_pulse}{a soundPulse object}
}
\description{
\itemize{
\item \code{duration}: the length of the pulse in ms
\item \code{peak_time}: the relative position of the peak in the pulse
\item \code{peak_frequency}: the frequency of the peak in kHz
\item \code{frequency_range}: the difference between highest and lowest frequency in the pulse (in kHz)
\item \code{start_frequency}: the relative distance between the peak frequency and the lowest frequency in the pulse. Defined as the difference devided by the frequency range.
\item \code{peak_amplitude}: the amplitude of the peak in dB. 0 dB is the median loudness
\item \code{PXXXX}: a set of pyramide values describing the shape of the pulse. The number of pyramide values depends on the total number of pulses and the dimensions of the shape
}
}
