% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconstruct.R
\name{reconstruct}
\alias{reconstruct}
\alias{reconstruct.list}
\alias{reconstruct.data.frame}
\alias{reconstruct.soundPulse}
\alias{reconstruct.soundPyramide}
\alias{reconstruct.soundCluster}
\title{Reconstruct pulses}
\usage{
reconstruct(x, ...)

\method{reconstruct}{list}(x, ...)

\method{reconstruct}{data.frame}(x, ...)

\method{reconstruct}{soundPulse}(x, ..., spectrogram)

\method{reconstruct}{soundPyramide}(x, ...)

\method{reconstruct}{soundCluster}(x, ...)
}
\arguments{
\item{x}{the object to reconstruct the pulse}

\item{...}{further arguments used by some methods}

\item{spectrogram}{the fingerprint of the selected spectrogram}
}
\description{
Reconstruct pulses
}
