% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound_spectrogram.R
\name{sound_spectrogram}
\alias{sound_spectrogram}
\title{Convert the output of read_wav to a spectrogram}
\usage{
sound_spectrogram(wav, window_ms = 1, overlap = 0.9)
}
\arguments{
\item{wav}{An object as created by \code{\link{sound_wav}}}

\item{window_ms}{The size of the window in microseconds. Default to 1.}

\item{overlap}{The overlap of two windows. Defaults to 0.9.}
}
\description{
Convert the output of read_wav to a spectrogram
}
\examples{
wav <- sound_wav(
  system.file("demo.wav", package = "soundcluster"),
  te_factor = 10,
  max_length = 0.1
)
sound_spectrogram(wav)
}
