% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound_wav.R
\name{sound_wav}
\alias{sound_wav}
\title{create a soundWav object}
\usage{
sound_wav(filename, timestamp, channel = c("left", "right"),
  te_factor = 1, sample_rate, values)
}
\arguments{
\item{filename}{The name of the file}

\item{timestamp}{the timestamp of the file}

\item{channel}{Select the left or the right channel}

\item{te_factor}{The factor to which the original sound was slowed down prior
to recording}

\item{sample_rate}{the realtime sample rate}

\item{values}{the recorded values}
}
\description{
create a soundWav object
}
