% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundCluster-class.R
\docType{class}
\name{soundCluster-class}
\alias{soundCluster-class}
\title{The soundCluster class}
\description{
Holds a cluster model.
}
\section{Slots}{

\describe{
\item{\code{Network}}{A self organising map}
\item{\code{Scaling}}{The scaling applied to the raw data}
}
}

