% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundDatabase-class.R
\docType{class}
\name{soundDatabase-class}
\alias{soundDatabase-class}
\title{The soundDatabase class}
\description{
It holds a connection to a soundDatabase
}
\section{Slots}{

\describe{
\item{\code{Connection}}{An open SQLiteConnection to the soundDatabase}
}
}

