% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundPulse-class.R
\docType{class}
\name{soundPulse-class}
\alias{soundPulse-class}
\title{The soundPulse class}
\description{
Holds pulses and their metadata.
}
\section{Slots}{

  \describe{
   \item{\code{Pulse}}{A data.frame with fingerprint, spectrogram, peak_time, start_time, end_time, peak_frequency, start_frequency, end_frequency, peak_amplitude, start_amplitude, select_amplitude and shape}
   \item{\code{Recording}}{A data.frame with fingerprint, filename, timestamp, sample_rate, te_factor, left_channel}
   \item{\code{Spectrogram}}{A data.frame with fingerprint, window_ms, window_n, overlap, recording}
  }
}

