% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundPyramid-class.R
\docType{class}
\name{soundPyramid-class}
\alias{soundPyramid-class}
\title{The soundPyramid class}
\description{
Holds pyramids of pulses and their metadata.
}
\section{Slots}{

\describe{
\item{\code{Pulse}}{A data.frame with \code{fingerprint} and its \code{spectrogram}}
\item{\code{Pyramid}}{A matrix with pyramid values}
\item{\code{Scaling}}{A matrix with \code{center} and \code{sd} for used to center each variable in \code{Pyramid}}
\item{\code{Recording}}{A data.frame with fingerprint, filename, timestamp, sample_rate, te_factor, left_channel}
\item{\code{Spectrogram}}{A data.frame with fingerprint, window_ms, overlap, recording}
}
}

