% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundPyramide-class.R
\docType{class}
\name{soundPyramide-class}
\alias{soundPyramide-class}
\title{The soundPyramide class}
\description{
Holds pyramides of pulses and their metadata.
}
\section{Slots}{

\describe{
\item{\code{Pulse}}{A data.frame with \code{fingerprint} and its \code{spectrogram}}
\item{\code{Pyramide}}{A matrix with pyramide values}
\item{\code{Scaling}}{A matrix with \code{center} and \code{sd} for used to center each variable in \code{Pyramide}}
\item{\code{Recording}}{A data.frame with fingerprint, filename, timestamp, sample_rate, te_factor, left_channel}
\item{\code{Spectrogram}}{A data.frame with fingerprint, window_ms, window_n, overlap, recording}
}
}

