% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpectrogram-class.R
\docType{class}
\name{soundSpectrogram-class}
\alias{soundSpectrogram-class}
\title{The soundSpectrogram class}
\description{
Holds a spectrogram and it's metadata.
}
\section{Slots}{

\describe{
\item{\code{SpecGram}}{The actual spectrogram}
\item{\code{Recording}}{A data.frame with fingerprint, filename, timestamp, sample_rate, te_factor, left_channel}
\item{\code{Spectrogram}}{A data.frame with fingerprint, window_ms, window_n, overlap, recording}
}
}

