% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundSpectrogramMeta-class.R
\docType{class}
\name{soundSpectrogramMeta-class}
\alias{soundSpectrogramMeta-class}
\title{The soundSpectrogramMeta class}
\description{
It holds metadata on sound recordings
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{A data.frame with fingerprint, filename, timestamp, sample_rate, te_factor, left_channel}
   \item{\code{Spectrogram}}{A data.frame with fingerprint, window_ms, window_n, overlap, recording}
  }
}

