% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundWav-class.R
\docType{class}
\name{soundWav-class}
\alias{soundWav-class}
\title{The soundWav class}
\description{
Holds the relevant content of wav files
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{The metadata on the recording}
   \item{\code{Values}}{The actual recorded values}
  }
}

