% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundWavMeta-class.R
\docType{class}
\name{soundWavMeta-class}
\alias{soundWavMeta-class}
\title{The soundWavMeta class}
\description{
It holds metadata on recordings
}
\section{Slots}{

  \describe{
   \item{\code{Recording}}{A data.frame with ID, Fingerprint, Filename, Timestamp, SampleRate, SampleRate, TEFactor, LeftChannel}
  }
}

