% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif.R
\name{vif}
\alias{vif}
\title{Calculate the Variance Inflation Factor of a data.frame}
\usage{
vif(x)
}
\arguments{
\item{x}{a data.frame with numeric variables}
}
\description{
Calculate the Variance Inflation Factor of a data.frame
}
