% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav2db.R
\name{wav2db}
\alias{wav2db}
\title{Extract pulses and store them in a soundcluster database}
\usage{
wav2db(db, path, recursive = TRUE, make, model, serial = NA_character_,
  te_factor = 1, channel = c("left", "right"), max_length = 30,
  window_ms = 1, overlap = 0.9, threshold_amplitude = 10,
  min_peak_amplitude = 30, dimensions = 32, existing = c("append",
  "skip"), frequency_range = c(10000, 130000))
}
\arguments{
\item{db}{A connection to a soundcluster database}

\item{path}{the name of a file or a directory}

\item{recursive}{logical.  Should the listing recurse into directories?}

\item{make}{the manufacturer of the device}

\item{model}{the model of the device}

\item{serial}{the optional serial number of the device}

\item{te_factor}{The factor to which the original sound was slowed down prior
to recording}

\item{channel}{Select the left or the right channel}

\item{max_length}{Maximum length of the recording to use in seconds. If the
recording is longer, the last part is ignored.}

\item{window_ms}{The size of the window in microseconds. Default to 1.}

\item{overlap}{The overlap of two windows. Defaults to 0.9.}

\item{threshold_amplitude}{relevant regions have an amplitude above the \code{threshold_amplitude}. Defaults to 10 dB.}

\item{min_peak_amplitude}{the maximum amplitude in a relevant region must be above \code{min_peak_amplitude}. Defaults to 30 dB.}

\item{dimensions}{the number of rows and columns used to resample the shape. Must be a single number and a power of 2. Will be altered to the next power of 2.}

\item{existing}{what to do with existing spectrograms. "append" will add new pulses. "skip" will skip the recording.}

\item{frequency_range}{the range of frequencies to use in Hz. Frequencies below the minimum or above the maximum are removed from the spectrogram. Defaults to 10000 (10 kHz) and 130000 (130 kHz).}
}
\description{
Extract pulses and store them in a soundcluster database
}
