% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wav2spectrogram.R
\name{wav2spectrogram}
\alias{wav2spectrogram}
\title{Convert the output of read_wav to a spectrogram}
\usage{
wav2spectrogram(wav, window_ms = 1, overlap = 0.9)
}
\arguments{
\item{wav}{An object as created by \code{\link{sound_wav}}}

\item{window_ms}{The size of the window in microseconds. Default to 1.}

\item{overlap}{The overlap of two windows. Defaults to 0.9.}
}
\description{
Convert the output of read_wav to a spectrogram
}
\examples{
wav <- sound_wav(
  system.file("demo.wav", package = "soundcluster"),
  channel = "right",
  te_factor = 10
)
wav2spectrogram(wav)
}
