% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sound_wav.R
\name{sound_wav}
\alias{sound_wav}
\title{create a soundWav object}
\usage{
sound_wav(filename, channel = c("left", "right"), te_factor = 1,
  max_length = 5)
}
\arguments{
\item{filename}{The name of the file}

\item{channel}{Select the left or the right channel}

\item{te_factor}{The factor to which the original sound was slowed down prior
to recording}

\item{max_length}{Maximum length of the recording to use in seconds. If the
recording is longer, the last part is ignored.}
}
\description{
create a soundWav object
}
\examples{
wav <- sound_wav(
  system.file("demo.wav", package = "soundcluster"),
  te_factor = 10,
  max_length = 0.1
)
}
