% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundCluster-class.R
\docType{class}
\name{soundCluster-class}
\alias{soundCluster-class}
\title{The soundCluster class}
\description{
Holds a cluster model.
}
\section{Slots}{

\describe{
\item{\code{Network}}{A self organising map}
\item{\code{Pulse}}{A data.frame with \code{fingerprint} and its \code{spectrogram}}
\item{\code{Scaling}}{A matrix with \code{center} and \code{sd} for used to center each variable in \code{Pyramid}}
\item{\code{Recording}}{A data.frame with fingerprint, filename, timestamp, sample_rate, te_factor, left_channel}
\item{\code{Spectrogram}}{A data.frame with fingerprint, window_ms, overlap, recording}
}
}

